fix: PrizeLotto 생성하여 당첨번호, 보너스번호 검증 책임
