+++ __tests__/lottoMachine.test.js
@@ -1,4 +1,4 @@
-import LottoMachine from "../src/domain/lottoMachine.js";
+import LottoMachine from "../src/domain/LottoMachine.js";
 
 describe("LottoMachine Class 테스트", () => {
     it("로또 머신에 금액을 넣을 시 1000원 이하면 안된다.", () => {
